Refactored googleAuth routes to routes folder and passport configuration to services folder
